add library for LCD formatting
